fix: decomposed TabbedModal compenents:
- ModalTabDesignPipeline
- ModalTabTestPipeline
- DialogConfirmCreatePipeline
- DialogPipelineCreationWarning
