+++ src/components/TabbedModal.tsx
@@ -1,61 +1,16 @@
 import React, { useState, useEffect } from 'react';
-import {
-  Modal,
-  Box,
-  Tabs,
-  Tab,
-  Typography,
-  IconButton,
-  Button,
-  TextField,
-  Divider,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogContentText,
-  DialogTitle,
-  Input,
-} from '@mui/material';
+import { Modal, Box, Tabs, Tab, Typography, IconButton } from '@mui/material';
 import { useTheme } from '@mui/material/styles';
 import CloseIcon from '@mui/icons-material/Close';
-import AddIcon from '@mui/icons-material/Add';
-import DeleteIcon from '@mui/icons-material/Delete';
-import ArrowUpwardIcon from '@mui/icons-material/ArrowUpward';
-import ArrowDownwardIcon from '@mui/icons-material/ArrowDownward';
-import Editor from '@monaco-editor/react';
-import Select, { ActionMeta, MultiValue, SingleValue } from 'react-select';
-import { getCustomStyles } from '../utils/getCustomStyles';
-import { SelectedOption } from '../types/SelectedOption';
 import { Processor } from '../types/processor';
 import { Schema, SchemaFormat } from '../types/schema';
-import { getProcessors, getTopicsAndSchemas } from '../utils/getEntities';
-import { postTestEvent } from '../utils/postTestEvent';
 import { FrontendPipeline, PipelineStep } from '../types/pipelines';
-import { v4 as uuidv4 } from 'uuid';
-import { postPipeline } from '../utils/postPipeline';
-import { getTestResults } from '../utils/getTestResults'
-
-interface TabPanelProps {
-  children?: React.ReactNode;
-  index: number;
-  value: number;
-}
-
-function TabPanel(props: TabPanelProps) {
-  const { children, value, index, ...other } = props;
-
-  return (
-    <div
-      role='tabpanel'
-      hidden={value !== index}
-      id={`tabpanel-${index}`}
-      aria-labelledby={`tab-${index}`}
-      {...other}
-    >
-      {value === index && <Box sx={{ p: 3 }}>{children}</Box>}
-    </div>
-  );
-}
+import { getProcessors, getTopicsAndSchemas } from '../utils/getEntities';
+import TabPanel from './TabPanel';
+import PipelineWarningDialog from './DialogPipelineCreationWarning';
+import PipelineConfirmationDialog from './DialogConfirmCreatePipeline';
+import DesignPipelineTab from './ModalTabDesignPipeline';
+import TestPipelineTab from './ModalTabTestPipeline';
 
 function a11yProps(index: number) {
   return {
@@ -93,8 +48,8 @@ const TabbedModal = ({ open, onClose, pipeline }: TabbedModalProps) => {
     null
   );
   const [steps, setSteps] = useState<[] | Processor[]>([]);
-  const [sourceTopic, setsourceTopic] = useState<string | null>(null);
-  const [targetTopic, settargetTopic] = useState<string | null>(null);
+  const [sourceTopic, setSourceTopic] = useState<string | null>(null);
+  const [targetTopic, setTargetTopic] = useState<string | null>(null);
   const [userCreatedPipeline, setUserCreatedPipeline] =
     useState<null | FrontendPipeline>(null);
 
@@ -131,44 +86,9 @@ const TabbedModal = ({ open, onClose, pipeline }: TabbedModalProps) => {
     loadProcessors();
   }, []);
 
-  const createEditableTestEvent = () => {
-    const request = async () => {
-      const result = await postTestEvent({
-        schema: incomingSchema as string,
-        format: selectedIncomingSchemaFormat as SchemaFormat,
-      });
-      setTestEvent(JSON.stringify(result, null, 2));
-    };
-    request();
-  };
-
   const handleTabChange = (_: React.SyntheticEvent, newValue: number) => {
     setTabValue(newValue);
   };
-
-  const handleTest = () => {
-    // const result = await simulateBackendProcessing(testEvent);
-    const request = async() => {
-      return await getTestResults(selectedIncomingSchemaFormat, testEvent, steps);
-    }
-    
-    
-    const result = request().then(res => {
-      console.log('result!!!!', res);
-      setTestResult(JSON.stringify(res.transformedMessage, null, 2));
-    });
-    console.log('Attempt to test result:', result)
-    // setTestResult(result);
-    return result
-  };
-
-  // const simulateBackendProcessing = (event: string) => {
-  //   return new Promise((resolve) => {
-  //     setTimeout(() => {
-  //       resolve(`Processed event: ${event}`);
-  //     }, 1000);
-  //   });
-  // };
 
   const handleClose = () => {
     setNewPipelineName('');
@@ -180,99 +100,6 @@ const TabbedModal = ({ open, onClose, pipeline }: TabbedModalProps) => {
     setdisplayedProcessorSelectOptions([]);
     setTabValue(0);
     onClose();
-  };
-
-  const handleMoveUp = (id: string) => {
-    const index = displayedProcessorSelectOptions.findIndex(
-      (item) => item.id === id
-    );
-    // console.log('Index:', index);
-    if (index === 0) return;
-    const newProcesses = [...displayedProcessorSelectOptions];
-    const temp = newProcesses[index - 1];
-    newProcesses[index - 1] = newProcesses[index];
-    newProcesses[index] = temp;
-    // console.log('New Processes:', newProcesses);
-    setdisplayedProcessorSelectOptions(newProcesses);
-  };
-
-  const handleMoveDown = (id: string) => {
-    const index = displayedProcessorSelectOptions.findIndex(
-      (item) => item.id === id
-    );
-    if (index === displayedProcessorSelectOptions.length - 1) return;
-    const newProcesses = [...displayedProcessorSelectOptions];
-    const temp = newProcesses[index + 1];
-    newProcesses[index + 1] = newProcesses[index];
-    newProcesses[index] = temp;
-    // console.log('New Processes:', newProcesses);
-    setdisplayedProcessorSelectOptions(newProcesses);
-  };
-
-  const handleDisplayNewProcessorSelect = (type: string, index: number) => {
-    const newProcessorSelectOptions = [...displayedProcessorSelectOptions];
-    const newProcessor = {
-      id: uuidv4(),
-      is_filter: type === 'filter',
-    };
-
-    newProcessorSelectOptions.splice(index + 1, 0, newProcessor);
-    setdisplayedProcessorSelectOptions(newProcessorSelectOptions);
-  };
-
-  const handleRemoveProcessorSelect = (id: number | string) => {
-    // console.log('Removing Processor with ID:', id, typeof id);
-    // console.log('Current steps:', steps);
-    // console.log('Current Processors:', displayedProcessorSelectOptions);
-    // console.log(
-    //   'Current Processor exists?',
-    //   displayedProcessorSelectOptions.some((item) => item.id === id)
-    // );
-    const index = displayedProcessorSelectOptions.findIndex(
-      (item) => item.id === id
-    );
-    const newSelectOptions = displayedProcessorSelectOptions
-      .slice(0, index)
-      .concat(displayedProcessorSelectOptions.slice(index + 1));
-
-    // console.log('New Processors:', newSelectOptions);
-    setdisplayedProcessorSelectOptions(newSelectOptions);
-    handleRemovePipelineStep(index);
-  };
-
-  const handleRemovePipelineStep = (index: number) => {
-    const newSteps = steps.slice(0, index).concat(steps.slice(index + 1));
-    // console.log('Previous steps:', steps);
-    // console.log('New Steps:', newSteps);
-    setSteps(newSteps);
-  };
-
-  const handleAddPipelineStep = (
-    selectedOption: SingleValue<SelectedOption> | MultiValue<SelectedOption>,
-    _: ActionMeta<SelectedOption>,
-    index: number
-  ) => {
-    // console.log('Adding Pipeline Step', selectedOption);
-    // console.log('Index:', index);
-    // console.log('Step to change', steps.at(index));
-    if (!selectedOption) return;
-
-    const processor = processorOptions.find(
-      (p) => p.processor_name === (selectedOption as SelectedOption).value
-    );
-    if (!processor) return;
-    let changeStep = steps.at(index);
-    if (changeStep) {
-      // console.log('New Step:', processor);
-      const newSteps = steps
-        .slice(0, index)
-        .concat([processor])
-        .concat(steps.slice(index + 1));
-
-      setSteps(newSteps);
-    } else {
-      setSteps([...steps, processor]);
-    }
   };
 
   const createPipelineObject = (): FrontendPipeline => {
@@ -288,17 +115,6 @@ const TabbedModal = ({ open, onClose, pipeline }: TabbedModalProps) => {
       steps,
     };
   };
-
-  const pipelineServerPost = () => {
-    const request = async () => {
-      if (!userCreatedPipeline) return;
-      return await postPipeline(userCreatedPipeline)
-    }
-    
-    const result = request();
-    console.log(result)
-    return result
-  }
 
   const handleCreatePipeline = () => {
     const pipeline = createPipelineObject();
@@ -366,428 +182,59 @@ const TabbedModal = ({ open, onClose, pipeline }: TabbedModalProps) => {
             <Tab label='Test' {...a11yProps(1)} />
             <Tab label='Versioning and Evolution' {...a11yProps(2)} />
           </Tabs>
-          <TabPanel value={tabValue} index={0}>
-            <Box
-              sx={{
-                display: 'flex',
-                flexDirection: 'Column',
-                gap: 0,
-              }}
-            >
-              <Typography>Name</Typography>
-              <Input
-                required
-                sx={{
-                  width: '300px',
-                  border: '1px solid #ccc', // Add a border
-                  borderRadius: '4px',
-                }}
-                onChange={(event) => setNewPipelineName(event.target.value)}
-              />
-            </Box>
-            <Divider sx={{ my: 2 }} />
-            <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
-              <Box sx={{ display: 'flex', gap: 2 }}>
-                <Box sx={{ flex: 1 }}>
-                  <Typography>Incoming Schema</Typography>
-                  <Select
-                    options={schemas.map((schema) => ({
-                      value: String(schema),
-                      label: String(schema),
-                    }))}
-                    isClearable
-                    styles={getCustomStyles(mode)}
-                    onChange={(
-                      selectedOption:
-                        | SingleValue<SelectedOption>
-                        | MultiValue<SelectedOption>,
-                      actionMeta: ActionMeta<SelectedOption>
-                    ) =>
-                      setIncomingSchema(
-                        selectedOption
-                          ? (selectedOption as SelectedOption).value
-                          : null
-                      )
-                    }
-                  />
-                </Box>
-                <Box sx={{ flex: 1 }}>
-                  <Typography>Incoming Schema Type</Typography>
-                  <Select
-                    options={['avro', 'json', 'protobuf'].map((schemaType) => ({
-                      value: schemaType,
-                      label: schemaType,
-                    }))}
-                    isClearable
-                    {...(incomingSchema ? { required: true } : {})}
-                    styles={getCustomStyles(mode)}
-                    onChange={(selectedOption) =>
-                      setSelectedIncomingSchemaFormat(
-                        (selectedOption as SelectedOption).value as SchemaFormat
-                      )
-                    }
-                  />
-                </Box>
-              </Box>{' '}
-              <Box>
-                <Typography>Source Topic</Typography>
-                <Select
-                  options={topics.map((topic) => ({
-                    value: topic,
-                    label: topic,
-                  }))}
-                  isClearable
-                  styles={getCustomStyles(mode)}
-                  onChange={(selectedOption) =>
-                    setsourceTopic(
-                      selectedOption
-                        ? (selectedOption as SelectedOption).value
-                        : null
-                    )
-                  }
-                />
-              </Box>{' '}
-              <Box>
-                <Typography>Target Topic</Typography>
-                <Select
-                  options={topics.map((topic) => ({
-                    value: topic,
-                    label: topic,
-                  }))}
-                  isClearable
-                  styles={getCustomStyles(mode)}
-                  onChange={(selectedOption) =>
-                    settargetTopic(
-                      selectedOption
-                        ? (selectedOption as SelectedOption).value
-                        : null
-                    )
-                  }
-                />
-              </Box>{' '}
-              <Box sx={{ display: 'flex', gap: 2 }}>
-                <Box sx={{ flex: 1 }}>
-                  <Typography>Outgoing Schema</Typography>
-                  <Select
-                    options={schemas.map((schema) => ({
-                      value: String(schema),
-                      label: String(schema),
-                    }))}
-                    isClearable
-                    styles={getCustomStyles(mode)}
-                    onChange={(selectedOption) =>
-                      setOutgoingSchema(
-                        selectedOption
-                          ? (selectedOption as SelectedOption).value
-                          : null
-                      )
-                    }
-                  />
-                </Box>
-                <Box sx={{ flex: 1 }}>
-                  <Typography>Topic On Fail</Typography>
-                  <Select
-                    options={topics.map((topic) => ({
-                      value: topic,
-                      label: topic,
-                    }))}
-                    isClearable
-                    styles={getCustomStyles(mode)}
-                    onChange={(selectedOption) =>
-                      setSchemaRedirectTopic(
-                        selectedOption
-                          ? (selectedOption as SelectedOption).value
-                          : null
-                      )
-                    }
-                  />
-                </Box>
-              </Box>
-              <Divider sx={{ my: 2 }} />
-              <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
-                {displayedProcessorSelectOptions.map((item, index) => (
-                  <Box
-                    key={item.id}
-                    sx={{
-                      display: 'flex',
-                      alignItems: 'flex-end',
-                      gap: 2,
-                    }}
-                  >
-                    {item.is_filter ? (
-                      <>
-                        <Box sx={{ flex: 1 }}>
-                          <Typography>
-                            Filter{' '}
-                            {displayedProcessorSelectOptions
-                              .filter((p) => p.is_filter)
-                              .indexOf(item) + 1}
-                          </Typography>
-                          <Select
-                            options={processorOptions
-                              .filter((p) => p.is_filter)
-                              .map((processor) => ({
-                                value: processor.processor_name,
-                                label: processor.processor_name,
-                              }))}
-                            onChange={(selectedOption, actionMeta) =>
-                              handleAddPipelineStep(
-                                selectedOption,
-                                actionMeta,
-                                index
-                              )
-                            }
-                            isClearable
-                            styles={getCustomStyles(mode)}
-                          />
-                        </Box>
-                        <Box sx={{ flex: 1 }}>
-                          <Typography>Topic On Fail</Typography>
-                          <Select
-                            options={topics.map((topic) => ({
-                              value: topic,
-                              label: topic,
-                            }))}
-                            // onChange={handleAddFilterTopic}
-                            isClearable
-                            styles={getCustomStyles(mode)}
-                          />
-                        </Box>
-                      </>
-                    ) : (
-                      <Box sx={{ flex: 1 }}>
-                        <Typography>
-                          Transformation{' '}
-                          {displayedProcessorSelectOptions
-                            .filter((p) => !p.is_filter)
-                            .indexOf(item) + 1}
-                        </Typography>
-                        <Select
-                          options={processorOptions
-                            .filter((p) => !p.is_filter)
-                            .map((processor) => ({
-                              value: processor.processor_name,
-                              label: processor.processor_name,
-                            }))}
-                          onChange={(selectedOption, actionMeta) =>
-                            handleAddPipelineStep(
-                              selectedOption,
-                              actionMeta,
-                              index
-                            )
-                          }
-                          isClearable
-                          styles={getCustomStyles(mode)}
-                        />
-                      </Box>
-                    )}
-                    <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-                      <IconButton
-                        onClick={() =>
-                          handleDisplayNewProcessorSelect(
-                            item.is_filter ? 'filter' : 'transformation',
-                            index
-                          )
-                        }
-                        sx={{ alignSelf: 'flex-end' }}
-                      >
-                        <AddIcon />
-                      </IconButton>
-                      <IconButton
-                        onClick={() => handleRemoveProcessorSelect(item.id)}
-                      >
-                        <DeleteIcon />
-                      </IconButton>
-                      <IconButton onClick={() => handleMoveUp(item.id)}>
-                        <ArrowUpwardIcon />
-                      </IconButton>
-                      <IconButton onClick={() => handleMoveDown(item.id)}>
-                        <ArrowDownwardIcon />
-                      </IconButton>
-                    </Box>
-                  </Box>
-                ))}
-              </Box>
-              <Box sx={{ display: 'flex', gap: 2 }}>
-                <Button
-                  onClick={() =>
-                    handleDisplayNewProcessorSelect(
-                      'transformation',
-                      displayedProcessorSelectOptions.length
-                    )
-                  }
-                  variant='contained'
-                  startIcon={<AddIcon />}
-                  sx={{ width: 'fit-content', alignSelf: 'flex-start' }}
-                >
-                  Add Transformation
-                </Button>
-                <Button
-                  onClick={() =>
-                    handleDisplayNewProcessorSelect(
-                      'filter',
-                      displayedProcessorSelectOptions.length
-                    )
-                  }
-                  variant='contained'
-                  startIcon={<AddIcon />}
-                  sx={{ width: 'fit-content', alignSelf: 'flex-start' }}
-                >
-                  Add Filter
-                </Button>
-              </Box>
-              <Button
-                onClick={handleCreatePipeline}
-                variant='contained'
-                color='secondary'
-                sx={{ width: 'fit-content', alignSelf: 'flex-start', mt: 2 }}
-              >
-                Update Pipeline
-              </Button>
-              <Button
-                onClick={() => pipelineServerPost()}
-                variant='contained'
-                color='secondary'
-                sx={{ width: 'fit-content', alignSelf: 'flex-start', mt: 2 }}
-              >
-                Create Pipeline
-              </Button>
-            </Box>
-          </TabPanel>
-          <TabPanel value={tabValue} index={1}>
-            <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
-              <Box sx={{ width: '200px' }}>
-                <Typography>Schema Type</Typography>
-                <Select
-                  options={[
-                    { value: 'json', label: 'JSON' },
-                    { value: 'avro', label: 'Avro' },
-                    { value: 'protobuf', label: 'Protobuf' },
-                  ]}
-                  value={{ value: schemaType, label: schemaType.toUpperCase() }}
-                  onChange={(selectedOption) =>
-                    setSchemaType(
-                      (selectedOption as SelectedOption).value as SchemaFormat
-                    )
-                  }
-                  styles={getCustomStyles(mode)}
-                />
-              </Box>
-              <Box>
-                <Button
-                  onClick={createEditableTestEvent}
-                  variant='contained'
-                  color='primary'
-                  sx={{ mr: 2 }}
-                >
-                  Generate Test Event
-                </Button>
-                <Button
-                  onClick={handleTest}
-                  variant='contained'
-                  color='secondary'
-                >
-                  Test
-                </Button>
-              </Box>
-              <Box sx={{ display: 'flex', gap: 2, height: '450px' }}>
-                <Box sx={{ flex: 1, display: 'flex', flexDirection: 'column' }}>
-                  <Typography variant='subtitle1'>
-                    Generated Editable Test Event:
-                  </Typography>
-                  <Box sx={{ flex: 1 }}>
-                    <Editor
-                      height='100%'
-                      language='json'
-                      value={testEvent}
-                      onChange={(value) => setTestEvent(value || '')}
-                    />
-                  </Box>
-                </Box>
-                <Box sx={{ flex: 1, display: 'flex', flexDirection: 'column' }}>
-                  <Typography variant='subtitle1'>Result:</Typography>
-                  <TextField
-                    multiline
-                    fullWidth
-                    rows={10}
-                    variant='outlined'
-                    value={testResult}
-                    InputProps={{
-                      readOnly: true,
-                      sx: { height: '100%' },
-                    }}
-                    sx={{ flex: 1 }}
-                  />
-                </Box>
-                <Box
-                  sx={{
-                    width: '200px',
-                    display: 'flex',
-                    flexDirection: 'column',
-                  }}
-                >
-                  <Typography variant='subtitle1'>
-                    Transformations and Filters:
-                  </Typography>
-                  <Box sx={{ mt: 2, flex: 1 }}>
-                    <Typography>Incoming Schema</Typography>
-                    <Typography>Transformation 1</Typography>
-                    <Typography>Filter 1</Typography>
-                    <Typography>Transformation 2</Typography>
-                    <Typography color='error'>Filter 2 (Error)</Typography>
-                    <Typography>Outgoing Schema</Typography>
-                  </Box>
-                </Box>
-              </Box>
-            </Box>
-          </TabPanel>
+          <DesignPipelineTab
+            tabValue={tabValue}
+            setNewPipelineName={setNewPipelineName}
+            schemas={schemas}
+            incomingSchema={incomingSchema}
+            setIncomingSchema={setIncomingSchema}
+            setSelectedIncomingSchemaFormat={setSelectedIncomingSchemaFormat}
+            setOutgoingSchema={setOutgoingSchema}
+            setSchemaRedirectTopic={setSchemaRedirectTopic}
+            mode={mode}
+            topics={topics}
+            setSourceTopic={setSourceTopic}
+            setTargetTopic={setTargetTopic}
+            processorOptions={processorOptions}
+            displayedProcessorSelectOptions={displayedProcessorSelectOptions}
+            setdisplayedProcessorSelectOptions={
+              setdisplayedProcessorSelectOptions
+            }
+            steps={steps}
+            setSteps={setSteps}
+            // pipelineServerPost={pipelineServerPost}
+            handleCreatePipeline={handleCreatePipeline}
+            userCreatedPipeline={userCreatedPipeline}
+          />
+          <TestPipelineTab
+            tabValue={tabValue}
+            incomingSchema={incomingSchema}
+            schemaType={schemaType}
+            setSchemaType={setSchemaType}
+            mode={mode}
+            selectedIncomingSchemaFormat={selectedIncomingSchemaFormat}
+            testEvent={testEvent}
+            setTestEvent={setTestEvent}
+            steps={steps}
+            testResult={testResult}
+            setTestResult={setTestResult}
+          />
           <TabPanel value={tabValue} index={2}>
             {/* Render Versioning and Evolution tab content here */}
             Versioning and Evolution Content
           </TabPanel>
         </Box>
       </Modal>
-      <Dialog
-        open={warningDialogOpen}
-        onClose={() => setWarningDialogOpen(false)}
-      >
-        <DialogTitle>Schema Validation Warning</DialogTitle>
-        <DialogContent>
-          <DialogContentText>
-            No outgoing schema selected. The pipeline will be processed without
-            schema validation. Is that okay?
-          </DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={() => setWarningDialogOpen(false)}>No</Button>
-          <Button
-            onClick={() => {
-              setWarningDialogOpen(false);
-              setConfirmDialogOpen(true);
-            }}
-            autoFocus
-          >
-            Yes
-          </Button>
-        </DialogActions>
-      </Dialog>
-      <Dialog
-        open={confirmDialogOpen}
-        onClose={() => setConfirmDialogOpen(false)}
-      >
-        <DialogTitle>Confirm Pipeline Creation</DialogTitle>
-        <DialogContent>
-          <DialogContentText>
-            Are you sure you want to create the pipeline?
-          </DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={() => setConfirmDialogOpen(false)}>No</Button>
-          <Button onClick={handleCreatePipeline} autoFocus>
-            Yes
-          </Button>
-        </DialogActions>
-      </Dialog>
+      <PipelineWarningDialog
+        warningDialogOpen={warningDialogOpen}
+        setWarningDialogOpen={setWarningDialogOpen}
+        setConfirmDialogOpen={setConfirmDialogOpen}
+      />
+      <PipelineConfirmationDialog
+        confirmDialogOpen={confirmDialogOpen}
+        setConfirmDialogOpen={setConfirmDialogOpen}
+        handleCreatePipeline={handleCreatePipeline}
+      />
     </>
   );
 };
