+++ src/utils/postTestEvent.ts
@@ -5,10 +5,11 @@ export const postTestEvent = async ({
   schema,
   format,
 }: Schema): Promise<JSON> => {
+  console.log('Test event schema:', schema);
   try {
     const result = await axios.post('http://localhost:3000/test_event', {
       format,
-      schema
+      schema,
     });
     return result.data;
   } catch (error) {
